+++ core/initializer/test_suite_launcher/test_suite_launcher.go
@@ -9,7 +9,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/docker/go-connections/nat"
-	"github.com/kurtosis-tech/kurtosis-client/golang/core_api_consts"
+	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_consts"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/docker_api/test_suite_container_mountpoints"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/docker_api/test_suite_env_vars"
 	test_suite_rpc_api_consts "github.com/kurtosis-tech/kurtosis-libs/golang/lib/rpc_api/rpc_api_consts"
@@ -271,7 +271,7 @@ func (launcher TestsuiteContainerLauncher) generateMetadataProvidingEnvVars() (m
 }
 
 func (launcher TestsuiteContainerLauncher) generateTestRunningEnvVars(kurtosisApiIp string) (map[string]string, error) {
-	kurtosisApiSocket := fmt.Sprintf("%v:%v", kurtosisApiIp, core_api_consts.ListenPort)
+	kurtosisApiSocket := fmt.Sprintf("%v:%v", kurtosisApiIp, kurtosis_core_rpc_api_consts.ListenPort)
 	result, err := launcher.generateTestSuiteEnvVars(kurtosisApiSocket)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred generating the test-running env vars")
@@ -285,7 +285,7 @@ Generates the map of environment variables needed to run a test suite container
 func (launcher TestsuiteContainerLauncher) generateTestSuiteEnvVars(kurtosisApiSocket string) (map[string]string, error) {
 	// TODO switch to the envVars requiring a visitor to hit, so we get them all
 	standardVars := map[string]string{
-		test_suite_env_vars.CustomParamsJson:        launcher.customParamsJson,
+		test_suite_env_vars.CustomParamsJsonEnvVar:        launcher.customParamsJson,
 		test_suite_env_vars.DebuggerPortEnvVar:      strconv.Itoa(portForDebuggersRunningOnTestsuite),
 		test_suite_env_vars.KurtosisApiSocketEnvVar: kurtosisApiSocket,
 		test_suite_env_vars.LogLevelEnvVar:          launcher.suiteLogLevel,